pyproject.toml
[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[project]
name = "papermap"
version = "2026.1.0"
description = "PaperMap is a Python package and CLI for creating ready-to-print paper maps."
authors = [{name = "Steven van de Graaf", email = "steven@vandegraaf.xyz"}]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["paper", "map", "topography", "osm", "openstreetmap"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]
requires-python = ">=3.10, <4"
dependencies = [
    "click",
    "click_default_group",
    "fpdf2",
    "Pillow",
    "requests",
]

[dependency-groups]
docs = [
    "cogapp",
    "furo",
    "myst-parser",
    "sphinx",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinxext-opengraph",
]
tests = [
    "pytest",
    "pytest-cov",
]
typing = [
    "mypy",
]
dev = [
    { include-group = "docs" },
    { include-group = "tests" },
    { include-group = "typing" },
    "prek",
]

[project.urls]
Documentation = "https://papermap.readthedocs.io/en/stable/"
Changelog = "https://papermap.readthedocs.io/en/stable/changelog.html"
"Source code" = "https://github.com/sgraaf/papermap"
Issues = "https://github.com/sgraaf/papermap/issues"

[project.scripts]
papermap = "papermap.cli:cli"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # missing-trailing-comma (flake8-commas), handled by the formatter
    "D100",  # undocumented-public-module (pydocstyle)
    "D102",  # undocumented-public-method (pydocstyle)
    "D105",  # undocumented-magic-method (pydocstyle)
    "D107",  # undocumented-public-init (pydocstyle)
    "E501",  # line-too-long (pycodestyle), handled by the formatter
    "PLC2401",  # non-ascii-name (Pylint)
]
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",  # unused-import (Pyflakes)
]
"docs/conf.py" = [
    "INP001",  # implicit-namespace-package (flake8-no-pep420)
]
"src/papermap/cli.py" = [
    "ANN401",  # any-type (flake8-annotations)
]
"src/papermap/utils.py" = [
    "N806",  # non-lowercase-variable-in-function (pep8-naming)
    "PLR2004",  # magic-value-comparison (Pylint)
]
"tests/**/test_*.py" = [
    "D",  # pydocstyle
    "PLR2004",  # magic-value-comparison (Pylint)
    "S101",  # assert (flake8-bandit)
]
"tests/conftest.py" = [
    "D401",  # non-imperative-mood (pydocstyle)
]

[tool.mypy]
python_version = "3.10"
